refactor(frontend): migrate shadcn ui to design-system/primitives
- Move 15 shadcn primitives into design-system/primitives, plus the
  composite data-table into design-system/components.
- Rewrite all consumer imports to use @/design-system barrel.
- Update components.json ui alias.
- Wire design-system/styles/theme.css into globals.css and load
  Manrope + Inter via next/font in the root layout.
- Expose shadcn compatibility vars (--color-background, --color-border,
  etc.) in theme.css's @theme inline block so Tailwind utilities like
  border-border and bg-background keep resolving after the globals.css
  rewrite.
