+++ frontend/components.json
@@ -15,7 +15,7 @@
   "aliases": {
     "components": "@/components",
     "utils": "@/lib/utils",
-    "ui": "@/components/ui",
+    "ui": "@/design-system/primitives",
     "lib": "@/lib",
     "hooks": "@/hooks"
   },
