+++ frontend/src/design-system/components/index.ts
@@ -1,2 +1,2 @@
-// Components built in Tasks 14–18.
-export {};
+export * from "./data-table";
+// status-badge, summary-card, fiscal-input, control-bar added in later tasks
