+++ frontend/src/design-system/primitives/table.tsx
