+++ frontend/src/design-system/styles/theme.css
@@ -93,6 +93,36 @@
   --color-ds-error-bg: var(--ds-error-bg);
   --color-ds-error-fg: var(--ds-error-fg);
 
+  /* Shadcn utility bridge — exposes the :root shadcn aliases as Tailwind color utilities
+   * (bg-background, text-foreground, border-border, etc.) so existing shadcn primitives
+   * and consumer classNames keep working unchanged. */
+  --color-background: var(--background);
+  --color-foreground: var(--foreground);
+  --color-card: var(--card);
+  --color-card-foreground: var(--card-foreground);
+  --color-popover: var(--popover);
+  --color-popover-foreground: var(--popover-foreground);
+  --color-primary: var(--primary);
+  --color-primary-foreground: var(--primary-foreground);
+  --color-secondary: var(--secondary);
+  --color-secondary-foreground: var(--secondary-foreground);
+  --color-muted: var(--muted);
+  --color-muted-foreground: var(--muted-foreground);
+  --color-accent: var(--accent);
+  --color-accent-foreground: var(--accent-foreground);
+  --color-destructive: var(--destructive);
+  --color-border: var(--border);
+  --color-input: var(--input);
+  --color-ring: var(--ring);
+  --color-sidebar: var(--sidebar);
+  --color-sidebar-foreground: var(--sidebar-foreground);
+  --color-sidebar-primary: var(--sidebar-primary);
+  --color-sidebar-primary-foreground: var(--sidebar-primary-foreground);
+  --color-sidebar-accent: var(--sidebar-accent);
+  --color-sidebar-accent-foreground: var(--sidebar-accent-foreground);
+  --color-sidebar-border: var(--sidebar-border);
+  --color-sidebar-ring: var(--sidebar-ring);
+
   /* Spacing utilities (additive to Tailwind's numeric scale) */
   --spacing-xs: 4px;
   --spacing-sm: 8px;
